# doc/swagger.yaml
openapi: 3.1.0
info:
  title: Readmaster.ai API
  description: API for Readmaster.ai, a web-based reading assessment platform.
  version: 0.1.0
paths:
...
  /api/v1/assessments:
    post:
      tags:
        - Assessments (Student)
      summary: Start New Assessment
      description: |-
        Allows a student to start a new assessment for a specified reading material.
        This endpoint creates an assessment record with a 'pending_audio' status.
      operationId: start_new_assessment_api_v1_assessments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAssessmentRequestDTO'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /api/v1/assessments/{assessment_id}/request-upload-url:
    post:
      tags:
        - Assessments (Student)
      summary: Request Assessment Audio Upload Url
      description: |-
        Requests a presigned URL for uploading the assessment audio file.
        The client should use this URL to PUT the audio file directly to the storage.
      operationId: request_assessment_audio_upload_url_api_v1_assessments__assessment_id__request_upload_url_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment for which to request an upload URL.
            title: Assessment Id
          description: The ID of the assessment for which to request an upload URL.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestUploadURLResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assessments/{assessment_id}/confirm-upload:
    post:
      tags:
        - Assessments (Student)
      summary: Confirm Assessment Audio Upload
      description: |-
        Confirms that the audio file for an assessment has been successfully uploaded
        by the client to the URL provided by /request-upload-url.
        This triggers the backend to update the assessment status and dispatch an AI processing task.
      operationId: confirm_assessment_audio_upload_api_v1_assessments__assessment_id__confirm_upload_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment for which upload is being confirmed.
            title: Assessment Id
          description: The ID of the assessment for which upload is being confirmed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadRequestDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmUploadResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assessments/{assessment_id}/quiz-answers:
    post:
      tags:
        - Assessments (Student)
      summary: Submit Quiz Answers For Assessment
      description: |-
        Allows a student to submit their answers to the quiz associated with an assessment.
        Calculates and stores the comprehension score.
      operationId: submit_quiz_answers_for_assessment_api_v1_assessments__assessment_id__quiz_answers_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment for which answers are being submitted.
            title: Assessment Id
          description: The ID of the assessment for which answers are being submitted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizSubmissionRequestDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSubmissionResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assessments/{assessment_id}/results:
    get:
      tags:
        - Assessments (Student)
      summary: Get Assessment Results
      description: |-
        Retrieves the detailed results of a specific assessment for the authenticated student.
        This includes AI analysis data, comprehension score, and a review of submitted quiz answers.
      operationId: get_assessment_results_api_v1_assessments__assessment_id__results_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment for which to retrieve results.
            title: Assessment Id
          description: The ID of the assessment for which to retrieve results.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResultDetailDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assessments/reading/{reading_id}:
    get:
      tags:
        - Assessments # Matches the updated router tag
      summary: List Assessments by Reading ID for Teachers/Parents
      description: |
        Allows authenticated Teachers and Parents to list all assessments for a specific reading material.
        The results are filtered based on the students/children they own or manage.
        - **Teachers**: Will see assessments for students enrolled in their classes. The `user_relationship_context` field will contain the class name.
        - **Parents**: Will see assessments for their linked children. The `user_relationship_context` field will typically indicate 'Your Child'.
        This endpoint is useful for tracking progress on assigned readings.
      operationId: list_assessments_by_reading_for_user_api_v1_assessments_reading__reading_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: reading_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Reading Id
          description: The ID of the reading material for which to list assessments.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
          description: Page number for pagination.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100 # Consistent with other paginated endpoints
            default: 20
            title: Size
          description: Number of items per page.
      responses:
        '200':
          description: Successful Response. Returns a paginated list of assessments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAssessmentListResponseDTO'
              examples:
                teacher_view:
                  summary: Example for a Teacher
                  value:
                    items:
                      - assessment_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        status: "completed"
                        assessment_date: "2024-01-15T10:30:00Z"
                        updated_at: "2024-01-15T11:00:00Z"
                        student:
                          student_id: "abf85f64-5717-4562-b3fc-2c963f66a123"
                          first_name: "John"
                          last_name: "Doe"
                          grade: "Grade 5"
                        reading:
                          reading_id: "cfd85f64-5717-4562-b3fc-2c963f66ab45"
                          title: "The Adventures of Tom Sawyer"
                        user_relationship_context: "English - Grade 5A"
                    page: 1
                    size: 1
                    total_count: 1
                parent_view:
                  summary: Example for a Parent
                  value:
                    items:
                      - assessment_id: "3fa85f64-5717-4562-b3fc-2c963f66afa7"
                        status: "pending_audio"
                        assessment_date: "2024-01-16T09:00:00Z"
                        updated_at: "2024-01-16T09:00:00Z"
                        student:
                          student_id: "bcf85f64-5717-4562-b3fc-2c963f66a456"
                          first_name: "Jane"
                          last_name: "Doe"
                          grade: null
                        reading:
                          reading_id: "cfd85f64-5717-4562-b3fc-2c963f66ab45"
                          title: "The Adventures of Tom Sawyer"
                        user_relationship_context: "Your Child"
                    page: 1
                    size: 1
                    total_count: 1
        '401':
          description: Unauthorized. Invalid or missing authentication token.
        '403':
          description: Forbidden. Authenticated user is not a Teacher or Parent, or does not have permissions for this reading's assessments.
        '404':
          description: Not Found. The specified `reading_id` does not exist.
        '422':
          description: Validation Error. Invalid parameter format (e.g., `reading_id` is not a valid UUID, or pagination parameters are out of bounds).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError' # Assumes HTTPValidationError is already defined
...
  /api/v1/parent/children/{child_student_id}/assessments/{assessment_id}/results:
    get:
      tags:
        - Parent - Child Monitoring
      summary: Parent Get Child Assessment Result Details
      description: Allows a parent to view detailed results of a specific assessment for one of their linked children.
      operationId: parent_get_child_assessment_result_details_api_v1_parent_children__child_student_id__assessments__assessment_id__results_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the child (student).
            title: Child Student Id
          description: The ID of the child (student).
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment.
            title: Assessment Id
          description: The ID of the assessment.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResultDetailDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssessmentResponseDTO:
      properties:
        assessment_id:
          type: string
          format: uuid
          title: Assessment Id
        student_id:
          type: string
          format: uuid
          title: Student Id
        reading_id:
          type: string
          format: uuid
          title: Reading Id
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        assessment_date:
          type: string
          format: date-time
          title: Assessment Date
        updated_at:
          type: string
          format: date-time
          title: Updated At
        audio_file_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Audio File Url
        audio_duration:
          anyOf:
            - type: integer
            - type: 'null'
          title: Audio Duration
          description: Duration of the audio in seconds.
        ai_raw_speech_to_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Ai Raw Speech To Text
          description: Raw speech-to-text output from AI processing.
        assigned_by_teacher_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Assigned By Teacher Id
        assigned_by_parent_id:
          type: string
          format: uuid
          nullable: true
          title: Assigned By Parent Id
      type: object
      required:
        - assessment_id
        - student_id
        - reading_id
        - status
        - assessment_date
        - updated_at
      title: AssessmentResponseDTO
      description: |-
        Standard DTO for representing an assessment in API responses.
        This can be used when an assessment is created or retrieved.
...
    ConfirmUploadRequestDTO:
      properties:
        blob_name:
          type: string
          title: Blob Name
          description: The unique blob name (path in storage) of the uploaded file.
      type: object
      required:
        - blob_name
      title: ConfirmUploadRequestDTO
      description: DTO for confirming that an audio file has been successfully uploaded to storage.
    ConfirmUploadResponseDTO:
      properties:
        assessment_id:
          type: string
          format: uuid
          title: Assessment Id
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        message:
          type: string
          title: Message
      type: object
      required:
        - assessment_id
        - status
        - message
      title: ConfirmUploadResponseDTO
      description: DTO for the response after confirming an upload and initiating processing.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QuizSubmissionRequestDTO:
      properties:
        answers:
          items:
            $ref: '#/components/schemas/QuizAnswerDTO'
          type: array
          title: Answers
          description: A list of answers submitted by the student.
      type: object
      required:
        - answers
      title: QuizSubmissionRequestDTO
      description: DTO for submitting a list of quiz answers for an assessment.
    QuizSubmissionResponseDTO:
      properties:
        assessment_id:
          type: string
          format: uuid
          title: Assessment Id
        comprehension_score:
          type: number
          title: Comprehension Score
          description: The calculated comprehension score (0.0 to 100.0).
        total_questions:
          type: integer
          title: Total Questions
          description: Total number of questions answered.
        correct_answers:
          type: integer
          title: Correct Answers
          description: Number of correctly answered questions.
        message:
          type: string
          title: Message
          default: Quiz answers submitted successfully.
      type: object
      required:
        - assessment_id
        - comprehension_score
        - total_questions
        - correct_answers
      title: QuizSubmissionResponseDTO
      description: DTO for the response after submitting quiz answers.
...
    RequestUploadURLResponseDTO:
      properties:
        upload_url:
          type: string
          title: Upload Url
          description: The presigned URL to which the client should upload the file.
        blob_name:
          type: string
          title: Blob Name
          description: The name of the blob (file path in storage) that will be created.
        required_fields:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Required Fields
          description: Any required fields or headers for the upload request.
      type: object
      required:
        - upload_url
        - blob_name
      title: RequestUploadURLResponseDTO
      description: DTO for the response when requesting a presigned URL for file upload.
    StartAssessmentRequestDTO:
      properties:
        reading_id:
          type: string
          format: uuid
          title: Reading Id
          description: The ID of the reading material for the assessment.
      type: object
      required:
        - reading_id
      title: StartAssessmentRequestDTO
      description: DTO for a student to request the start of a new assessment.
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login